exo3_projet_final/requirements.txt
napalm>=4.1.0
netmiko>=4.1.2
python-nmap>=0.7.1
paramiko>=3.3.0
requests>=2.28.0
pysnmp>=4.4.12
pandas>=1.5.0
matplotlib>=3.6.0
openpyxl>=3.0.10
dash>=2.9.0
plotly>=5.13.0
pyyaml>=6.0
ipaddress>=1.0.23
